Reviewed some TODO's by Peter:
TODO: [PW] Revise use of MUST, SHOULD, COULD, etc. It seems SHOULD and
SHOULD NOT is used when MUST/MUST NOT should be used. RECOMMENDED
should be SHOULD.

resolution by YD: revised some of them. Others are intentional.

TODO: Should be in section Implementation Notes. This section is for
the protocol schema, not the schemas used for compression.

resolution by YD: moved to implementation notes in the previous
commit.

TODO: [PW] Why not efficient? In terms of total amount of
communication performed during a device's lifetime, its a very very
small fraction (insignificant).  This is a subjective opinion, termed
in absolute phraseology.

resolution by YD: Some phrase are added to describe the problem of
diversity.